💥 (exchange): Merge Lobby into Exchange
Exchange Protocol now manages Discovery of Peers locally and Globally
