--- proto/protocols/lobby.proto
@@ -1,12 +1,0 @@
-syntax = "proto3";
-
-package sonr.protocols.lobby;
-import "proto/common/core.proto";
-option go_package = "github.com/sonr-io/core/pkg/lobby";
-
-// PublishEvent is message passed from Peer in Lobby Protocol
-message LobbyMessage {
-  sonr.core.Peer peer = 1;     // Users Peer Data
-  optional string message = 2; // Message to be published
-  sonr.core.Metadata metadata = 3; // Metadata
-}
